Revert typescript bump.
It seems to be building broken typescript. Specifically, the error I'm seeing is

```
node_modules/mobx-quick-tree/dist/src/index.d.ts:15:250 - error TS1005: '?' expected.
```

And that line refers to the generated type for `types.compose`, which is quite
large. Although I tested this and it seemed to work, I must have not been using
the built version locally.