Add class models, a faster API for readonly instances
This introduces a new API for defining models that allows us to construct the read only instances much faster. Without any optimization yet, our primitive benchmark shows a 2.5x speedup for this new API over the existing readonly API:

```
claw ~/C/mobx-quick-tree (class-models) ➜  yarn x spec/bench/cross-framework.ts
yarn run v1.22.10
$ ts-node --transpile-only spec/bench/cross-framework.ts
mobx-state-tree x 4,618 ops/sec ±12.62% (80 runs sampled)
mobx-quick-tree types.model x 98,964 ops/sec ±3.50% (91 runs sampled)
mobx-quick-tree ClassModel x 261,246 ops/sec ±0.22% (92 runs sampled)
Fastest is mobx-quick-tree ClassModel
Slowest is mobx-state-tree
```

The central premise is that `mobx-state-tree`'s API for defining models collects functions that add views and actions in such a way that each view-adder function or action-adder function has to be run once *per* instance. If you have a model like:

```typescript
const Todo = types.model("Todo", {
  name: types.string,
  done: types.boolean
}).actions(self => {
  setDone(done: boolean) {
    self.done = done;
  }
});
```

each time you instantiate a `Todo` with `Todo.create` or `Todo.createReadOnly`, that `self => ({ setDone() {} })` block has to run once to create a new set of actions which close over a new value of self. There's no magic behind the scenes that only runs that function once and re-uses the output between instances -- it's instead re-executed on instantiation every time. NIce API, terrible for performance.

Instead, we should leverage JavaScripts existing mechanisms for defining what is the same about a bunch of objects once instead of over and over: the prototype chain! Classes (which in JS are just sugar over prototypes) allow us to define something like the `setDone` function once on a prototype, and then create an instance which has that as it's prototype, and not need to re-create `setDone` per instance. This whole PR is about defining a class-based API that leverages this built-in performant approach in JS to power read only instances, while still allowing us to generate the observable instances we need sometimes. It's an inversion from the previous approach to MQT: before, we adapted the MST API to give us readonly instances while conforming to the existing API, but now, we define a new API with a higher performance ceiling, and adapt that API to the slower thing under the hood.

The implementation of this is scary to be honest. `types.model` was clearly modelled after classes, so the grain of class definitions and MST type defintions match, but getting it to work at runtime and typetime requires lots of fancy stuff. The first tricky part is the types. I think it's key that a new, class based API for MQT is incrementally adoptable so that we can roll it out over time and build confidence, and that means that the existing `types.model` types and MQT class based types need to interoperate. After much fiddling, I think I accomplished this, but it was hard to get `IType` and `IClassModelType` working in harmony. See the note in `IClassModelType` for more info.

Also tricky is the runtime derivation of the observable class from the readonly class. The decorators give us enough structure to hang off of, but ES2020/TS decorators have two key limitations: they can't muck with properties until they are defined, and TS decorators can't change the type. This matters most for volatiles and async actions.

Let's review the requirements:

 - Async actions in MST and plain old mobx are really several chunks of sync action chained together, which some kind of hook system to ensure that each sync bit is wrapped in mobx action wrapper goodness, and that inbetween while awaiting something, the wrapper isn't active. mobx/MST need this hook point to enforce actions are actually the only ones changing data.
 - We want to define async actions on classes

And then the limitations:

 - The only hook system that the mobx or MST folks have come up with that is ergonomic is having userland code build async actions using generators instead of the normal async/await. This is so they have their hook point for every `yield`, where they can be like "oh, a promise!" and disable the action bits for the duration of the promise, and re-active them when it resolves and they pass the value back in to the caller. Generators allow userland to intercept each call to what would be `await` in both node and the browser.
 - *Inside* the async action you use `yield somePromise`, but calling the action still returns a promise, so you still do `await someNode.someAsyncAction()`. This is to stay as normal as possible outside of MST action bodies and not make the whole system use generators. So, to adapt the generator into a normally typed async function, users need to wrap their generator functions in `flow()`.
 - There's no way to automatically adapt an async function into a generator function for MST. If there was, MST would use it already and not require folks to write generators. This means the code on the MQT class model *has* to be in generator form.
 - TypeScript decorators can't change the type of the property they are decorating, so they can't automatically apply the `flow` expression to a prototype level generator function and get the type turned into a nice promise returning type. If you have a class body like this:

```typescript

class Foo {

  @flow
  *someAsyncAction() {
  }
```

the return type of that function is locked in as a generator, and `@flow` can't change it. It can do stuff to it at *runtime*, but it can't mutate the type time type. That means that `await someFoo.someAsyncAction()` won't typecheck, as TypeScript will be expecting it to return a generator regardless of what runtime trickery the decorator gets up to. Sad.

So, we need to do the adaptation of the generator function into a promise-returning function for nice caller types with some sort of expression. That's what the `flow()` helper from MST does.

 - To assign a `flow(function * () {})` expression to a property on a class, it becomes an "instance property" instead of a prototype property. That is:

```typescript
class Foo {

  someAsyncAction = flow(function * () {
    yield sleep(100)
  });

}
```

desugars into:

```javascript
class Foo {

  constructor() {
    this.someAsyncAction = flow(function * () {
      yield sleep(100)
    });
  }

}
```

It'd be far more performant in general to have this somehow do a `Foo.prototype.whatever = flow(...)`, but alas, thats not how ES6 expression assignments work. This is a major issue for us though, because this desugaring means that there is nothing to look at to get the flow definition at class definition time. For things defined on the prototype, we can get a handle on them easily at class definition time to send them into mobx land for making the observable instance, but, for these things which aren't set until the constructor, we can't see their value until the constructor runs! After much brain wracking I couldn't find a away around this -- I had to go with a major hack which is to actually construct an instance, let the constructor run, and then pull the resulting expression off of the constructed instance to get a handle on the flow implementation to pass into MST. Yikes.