Flip inference order of InstanceWithoutSTNTypeForType back to checking IType first
Not exactly sure why, but this change during code review of the class models branch caused typechecking failure Gadget side with some of the most advanced generic-over-an-mqt-model types. We may want to flip this back around when we convert some of those types to be class model only, but for now, I think I am fine with this being a little less future proof but inference still working fine.