Fix typescript types for snapshotting optional types
This fixes a big issue after adding class models where ITypes wrapping class models didn't correctly instantiate. It all stems from IStateTreeNode -- the existing types use it to decorate any ole object with the `$type` it is of. This facilitates grabbing the type at runtime for API functions like `getType`, and for grabbing the type at type time to power type safety in the API for things like `getSnapshot` etc.

IStateTreeNode is thusly key, and it used to be that everything that was intantiated could be wrapped in an `IStateTreeNode` wrapper pretty easily. We need to keep doing this for IMaybeType and IReferenceType and the other ITypes which wrap some other inner type, but we don't want to do it to class models, which define their own `$type` property, and break things if extended with `& { [$type]: ... }`.

Making IStateTreeNode not extend the type if the type already has a `$type` fixes static typing. Jesus.