Allow union type instantiation to be faster by specifying a discriminator property
Gadget side, we instantiate a lot of unions that have many elements. Both MQT (or MST) have to take an incoming snapshot and figure out which element of the union is the right one to instantiate and return for the union. Both of them have a pretty naive algorithm for doing this: take each element and check the snapshot against it, returning the first union element which does. This means a LOT of `type.is` checks, where each type in the union is compared against the snapshot. These type.is checks basically do a full instantiation under the hood -- there's no fancy fast path for `.is` for a lot of types, there's just running the runtime code that would boot an instance, and seeing if it fails.

This works in the general case but has gotten really slow in Gadget, specifically for the FieldConfig enum. It has a lot of elements, and each element's type is complicated-ish such that it takes a little bit of time to check with `.is`.

Our snapshots for this type have a really easy to get at `"type": "StringConfig"` property on them though that allows us to discriminate between the union elements super easily. We added this a long time ago for typescript discriminated unions at type type, but, I think we can use this at runtime too for a much faster path for instantiation. MST already has support for a custom union type "dispatcher", that returns the right type to use for instantiation given a snapshot. This PR adds support for that, and then adds a macro of sorts on top that does type dispatching using a discriminating property. We expect all the types in the union to have a value for this property we can inspect at the time the union is defined, and then we build a map from the value of the discriminator to the type. When a snapshot comes in, we look up the value of the snapshot's discriminator property within the map.