Move the parent-setting bit into the instantiate method of each type
The `setParent` helper function was at the top of the large root instantiation benchmark for two reasons. One, we called it on a lot of things that weren't objects, like the primitive strings and numbers and dates in the tree, and two, Reflect.defineProperty is quite a bit slower than normal property access. Instead of blindly calling it from all the places where we construct a child, let's pass it down to the child constructor so that we can