Compile an optimized instantiation function for each MQT model that unrolls the property setup loop
Class model instantiation is a big bottleneck Gadget side still, so we're super motivated to make it as fast as it can be. Profiling has shown that the specific function that loops over the properties for a model and sets them on the instance is actually pretty slow. This is a challenging thing for the JIT to optimize because the functions we're going to call are super dynamic -- they're in an object at a weird variety of keys. Also sucks for branch prediction.

So, we deploy the heaviest of JS perf weapons: function compilation. Instead of a function which uses metadata dynamically to call the sub-type instantiation functions, we compile a new function that does the same thing, but explicitly in order. No loops or conditionals or hard-to-predict jumps into other instantiation functions, especially if the other function is super duper simple.