Eval a fast property memoizer to microoptimize property access for cached views
This implements a better performing getter function for readonly instances. Before this, we used the same definition of the getter function for all views, which ended up megamorphic, because it accessed a very wide variety of properties for every instance! Instead, this evals a monomorphic getter for each one. I also changed the object that stores the memos to have a fixed shape from birth by evaling it out as well.

I also changed us to use the object itself to store all the memoized values, instead of two other tracker objects. We were previously using the trackers in order to implement memoization of undefined and null correctly. Now, we set up symbol properties to store all the memos, but on the prototype of the class model. This means we don't spend any time setting them up during construction, but the JS VM knows the shape of the memos ahead of time! Less memory and same speed woop woop! Then, when memoizing, we can stash the value under that symbol key, which means fewer property accesses and lookups to fetch it.