Eval a fast property memoizer to microoptimize property access for cached views
This implements a better performing getter function for readonly instances. Before this, we used the same definition of the getter function for all views, which ended up megamorphic, because it accessed a very wide variety of properties for every instance! Instead, this evals a monomorphic getter for each one. I also changed the object that stores the memos to have a fixed shape from birth by evaling it out as well.

I also changed us to use one object to store all the memoized values, instead of two. We were previously using two in order to implement memoization of undefined and null correctly, but with the new strategy to initialize an object with slots for every memo from the start, we can populate it with a `$notYetMemoized` symbol that indicates if we have memoized or not.

And finally -- the object that stores the memos itself also has a dynamic shape, so it's also optimized! We create a new dynamic class per ClassModel that has all the properties we might memoize on it's prototype, which makes it fast to instantiate but have a stable shape.