Increase `initCount` for our recorded benchmarks.
This values is the number of times to run a benchmark before recording values.
It defaults to 1, but we want to have it be a high enough number to be more
likely to warm up the JIT so that benchmarks will likely be a bit fairer across
branches.

I'm not sure what a good value is, so this is very "finger in the wind".