Ensure types.maybe(types.reference()) types are resolved late after construction
This fixes a bug in the fast instantiator code where valid references would throw a `can't resolve reference` error due to the instantiation order within MQT. We were accidentally eagerly resolving some references when constructing a tree, instead of late resolving them at the end like we normally do. This only happened for wrapped reference types like a `maybe(reference())` or a `late(reference())` because of an instanceof check instead of a deeper `isReferenceType` check.

Fixes #78, which was an overly broad bug report as it turns out -- class models and normal models correctly late resolve references, regardless of property definition order before this change, it's just when you wrap the reference in a maybe that things got wonky.

There's one important little catch as well, which is that unions of
references are `isReferenceType()` in MST. We need to carefully handle
these, and not assume all union elements are references if one is. For
these rare but complicated cases, we fall back to the normal
instantiation path that is a bit slower, but run it in a callback at the
end of construction so that any reference members of the union are
running at the right time.

Woop woop!