Only add models that might need dereferencing to the reference cache
Previously, we were tracking all instances of all types in the reference cache when instantiating a root. We don't really need all references -- we only need the instances that we know might need de-referencing at some point. This implements a type-time, memoized traversal to check which types ever have incoming references, and only tracks instances in the reference cache if so.

[no-changelog-required]